refactor: resolve widget positions in the install converter only
Every other optional manifest field takes its default in the converter, and
stamping positions at build time would freeze the current default into every
published artifact.
