+++ packages/twenty-server/src/engine/core-modules/application/application-manifest/converters/from-page-layout-widget-manifest-to-universal-flat-page-layout-widget.util.ts
@@ -1,12 +1,49 @@
+import { type PageLayoutWidgetManifest } from 'twenty-shared/application';
+import { DEFAULT_WIDGET_SIZE } from 'twenty-shared/constants';
 import {
-  getPageLayoutWidgetManifestGridPosition,
-  getPageLayoutWidgetManifestPosition,
-  type PageLayoutWidgetManifest,
-} from 'twenty-shared/application';
-import { type PageLayoutTabLayoutMode } from 'twenty-shared/types';
+  PageLayoutTabLayoutMode,
+  type GridPosition,
+  type PageLayoutWidgetPosition,
+} from 'twenty-shared/types';
+import { isDefined } from 'twenty-shared/utils';
 
 import { type WidgetType } from 'src/engine/metadata-modules/page-layout-widget/enums/widget-type.enum';
 import { type UniversalFlatPageLayoutWidget } from 'src/engine/workspace-manager/workspace-migration/universal-flat-entity/types/universal-flat-page-layout-widget.type';
+
+// Widget placement is declared through the widgets array order and the tab layout mode:
+// manifests published before the position union existed carry none at all
+const getPageLayoutWidgetPosition = ({
+  pageLayoutWidgetManifest,
+  pageLayoutTabLayoutMode,
+  gridPosition,
+  widgetIndex,
+}: {
+  pageLayoutWidgetManifest: PageLayoutWidgetManifest;
+  pageLayoutTabLayoutMode: PageLayoutTabLayoutMode;
+  gridPosition: GridPosition;
+  widgetIndex: number;
+}): PageLayoutWidgetPosition => {
+  if (isDefined(pageLayoutWidgetManifest.position)) {
+    return pageLayoutWidgetManifest.position;
+  }
+
+  switch (pageLayoutTabLayoutMode) {
+    case PageLayoutTabLayoutMode.VERTICAL_LIST:
+      return {
+        layoutMode: PageLayoutTabLayoutMode.VERTICAL_LIST,
+        index: widgetIndex,
+      };
+    case PageLayoutTabLayoutMode.CANVAS:
+      return {
+        layoutMode: PageLayoutTabLayoutMode.CANVAS,
+      };
+    case PageLayoutTabLayoutMode.GRID:
+      return {
+        layoutMode: PageLayoutTabLayoutMode.GRID,
+        ...gridPosition,
+      };
+  }
+};
 
 export const fromPageLayoutWidgetManifestToUniversalFlatPageLayoutWidget = ({
   pageLayoutWidgetManifest,
@@ -23,6 +60,13 @@ export const fromPageLayoutWidgetManifestToUniversalFlatPageLayoutWidget = ({
   applicationUniversalIdentifier: string;
   now: string;
 }): UniversalFlatPageLayoutWidget => {
+  const gridPosition = pageLayoutWidgetManifest.gridPosition ?? {
+    row: 0,
+    column: 0,
+    rowSpan: DEFAULT_WIDGET_SIZE.default.h,
+    columnSpan: DEFAULT_WIDGET_SIZE.default.w,
+  };
+
   return {
     universalIdentifier: pageLayoutWidgetManifest.universalIdentifier,
     applicationUniversalIdentifier,
@@ -34,14 +78,11 @@ export const fromPageLayoutWidgetManifestToUniversalFlatPageLayoutWidget = ({
     objectMetadataUniversalIdentifier:
       pageLayoutWidgetManifest.objectUniversalIdentifier ?? null,
     conditionalDisplay: pageLayoutWidgetManifest.conditionalDisplay ?? null,
-    gridPosition: getPageLayoutWidgetManifestGridPosition(
-      pageLayoutWidgetManifest,
-    ),
-    // The build resolves the position, but manifests published by earlier SDK versions
-    // carry none: resolving it here as well heals those apps on their next install
-    position: getPageLayoutWidgetManifestPosition({
+    gridPosition,
+    position: getPageLayoutWidgetPosition({
       pageLayoutWidgetManifest,
       pageLayoutTabLayoutMode,
+      gridPosition,
       widgetIndex,
     }),
     universalConfiguration:
