+++ packages/twenty-server/src/engine/core-modules/application/application-manifest/converters/get-page-layout-tab-manifest-layout-mode.util.ts
@@ -1,8 +1,8 @@
 import {
   type PageLayoutManifest,
   type PageLayoutTabManifest,
-} from '@/application/pageLayoutManifestType';
-import { PageLayoutTabLayoutMode, PageLayoutType } from '@/types';
+} from 'twenty-shared/application';
+import { PageLayoutTabLayoutMode, PageLayoutType } from 'twenty-shared/types';
 
 export const getPageLayoutTabManifestLayoutMode = ({
   pageLayoutTabManifest,
