+++ packages/twenty-shared/src/application/index.ts
@@ -149,9 +149,6 @@ export {
   serializeApplicationVariableValue,
   deserializeApplicationVariableValue,
 } from './utils/applicationVariableValueSerialization';
-export { getPageLayoutTabManifestLayoutMode } from './utils/get-page-layout-tab-manifest-layout-mode.util';
-export { getPageLayoutWidgetManifestGridPosition } from './utils/get-page-layout-widget-manifest-grid-position.util';
-export { getPageLayoutWidgetManifestPosition } from './utils/get-page-layout-widget-manifest-position.util';
 export type {
   ViewManifestFilterValue,
   ViewFieldManifest,
